Move resourcecachingclient to pkg to be importable from other projects
